+++ pyroomacoustics/bss/auxiva.py
@@ -101,9 +101,11 @@ def auxiva(X, n_src=None, n_iter=20, proj_back=True, W0=None,
         G_r[:,:] = f_contrast['df'](r, *f_contrast_args) / r  # shape (n_frames, n_src)
 
         # Compute Auxiliary Variable
-        for f in range(n_freq):
-            for s in range(n_src):
-                V[f,s,:,:] =  (np.dot(G_r[None,:,s] * X[:,f,:].T, np.conj(X[:,f,:]))) / X.shape[0]
+        V = np.mean(
+                (X[:,:,None,:,None] * G_r[:,None,:,None,None])
+                * np.conj(X[:,:,None,None,:]),
+                axis=0,
+                )
 
         # Update now the demixing matrix
         for s in range(n_src):