+++ pyroomacoustics/soundsource.py
@@ -78,7 +78,8 @@ class SoundSource(object):
         self.max_order = np.max(self.orders)
 
 
-    def add_signal(signal):
+    def add_signal(self, signal):
+        ''' Sets ``SoundSource.signal`` attribute '''
 
         self.signal = signal