+++ CHANGELOG.rst
@@ -23,10 +23,12 @@ Changed
 
 - Limits imports of matplotlib to within plotting functions so that the
   matplotlib backend can still be changed, even after importing pyroomacoustics
+- Better Vectorization of the computations in ``pyroomacoustics.bss.auxiva``
 
 Bugfix
 ~~~~~~
 
+- Corrects a typo in ``pyroomacoustics.SoundSource.add_signal``
 - Corrects a bug in the update of the demixing matrix in ``pyroomacoustics.bss.auxiva``
 - Corrects invalid memory access in the ``pyroomacoustics.build_rir`` cython accelerator
   and adds a unit test that checks the cython code output is correct